metadata/requirements.txt
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
crypto==1.4.1
dotenv==0.9.9
et_xmlfile==2.0.0
GeoAlchemy2==0.17.1
greenlet==3.2.2
idna==3.10
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Naked==0.1.32
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
shapely==2.1.1
shellescape==3.8.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
